Оформила модуль map
