+++ js/map.js
@@ -1,8 +1,9 @@
 import {toggleFormsCondition, initialCoordinates} from './user-form.js';
-import {createNewCard} from './card.js';
+import {addMarkerTooltip, createNewCard} from './card.js';
 // import {initialCoordinates} from './user-form.js';
 
 export let map;
+export let markerGroup;
 
 export const createOfferMarker = (mainPinMarker) => {
   const offerPinIcon = L.icon({
@@ -37,7 +38,7 @@ export const initMap = (canvas, mainPinMarker) => {
       lat: initialCoordinates.lat,
       lng: initialCoordinates.lng,
     }, 13);
-
+  markerGroup = L.layerGroup().addTo(map);
   L.tileLayer(
     'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
     {
@@ -48,4 +49,28 @@ export const initMap = (canvas, mainPinMarker) => {
   mainPinMarker.addTo(map);
 };
 
-
+export const showSingleMarker = (offer) => {
+  const sameOfferIcon =  L.icon({
+    iconUrl: 'img/pin.svg',
+    iconSize: [40, 40],
+    iconAnchor: [20, 40],
+  });
+  const sameOfferMarker = L.marker({
+    lat: offer.location.lat,
+    lng: offer.location.lng,
+  },
+  {
+    icon: sameOfferIcon,
+  },
+  );
+  sameOfferMarker
+    .addTo(markerGroup)
+    .bindPopup(addMarkerTooltip(offer),
+      {
+        keepInView: true,
+      },
+    );
+};
+export const clearLayer = () => {
+  markerGroup.clearLayers();
+};
